requirements.txt
streamlit
pandas
matplotlib
numpy
plotly
scipy
biopython